// web/src/api/generated/model/inboxQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface InboxQuestion {
  id: string;
  reference: string;
  name: string;
  student_number: string | null;
  class_group: string | null;
  question: string;
  created_at: string;
  reply: string | null;
  replied_at: string | null;
  replied_by: string | null;
}